feat: implements the start to atomic design.
