+++ README.md
@@ -1,1 +1,64 @@
-# pomodoro-app
+# ⏱️ Pomodoro-app
+
+**pomodoro-app** é uma aplicação web desenvolvida com as principais tecnologias
+de frontend baseada na técnica de produtividade **Pomodoro**, que ajuda você a
+gerenciar melhor seu tempo de estudo ou trabalho.
+
+---
+
+## 🚀 Tecnologias Utilizadas
+
+- [TypeScript](https://www.typescriptlang.org/)
+- [Vite](https://vitejs.dev/)
+- [React](https://react.dev/)
+
+---
+
+## 📦 Instalação
+
+### Clone o repositório
+
+```bash
+git clone https://github.com/matheus-nicolau/pomodoro-app.git
+cd pomodoro-app
+```
+
+### Instale as dependências
+
+```bash
+npm install
+# ou
+yarn
+
+```
+
+### Inicie o servidor de desenvolvimento
+
+```bash
+npm run dev
+# ou
+yarn dev
+
+```
+
+```
+| Comando         | Descrição                          |
+| --------------- | ---------------------------------- |
+| npm run dev     | Inicia a aplicação em modo dev     |
+| npm run build   | Gera o build de produção           |
+| npm run preview | Visualiza o build localmente       |
+| npm run lint    | (se configurado) Lint nos arquivos |
+
+```
+
+⏳ Temporizador personalizável (pomodoro, pausa curta, pausa longa)
+
+🔔 Alarme ao término do ciclo
+
+🌗 Modo claro/escuro (opcional)
+
+🔄 Contagem automática de ciclos (opcional)
+
+---
+
+⚙️ `Desenvolvido por Matheus Nicolau`
