+++ src/App.tsx
@@ -1,13 +1,5 @@
-import './styles/global.css';
+import './global/global.css';
 import { createRoot } from 'react-dom/client';
-import { Hello } from './components/hello/Hello';
+import { Hello } from './components/organisms/hello/Hello';
 
-createRoot(document.getElementById('root')!).render(
-  <div className='container-fluid'>
-    <div className='container'>
-      <div className='content'>
-        <Hello />
-      </div>
-    </div>
-  </div>,
-);
+createRoot(document.getElementById('root')!).render(<Hello />);
