+++ src/components/atoms/menuMainButton/MenuMainButton.tsx
@@ -1,0 +1,14 @@
+import { TimerIcon } from 'lucide-react';
+
+function MenuMainButton() {
+  return (
+    <>
+      <a href='#'>
+        <TimerIcon />
+        Clique
+      </a>
+    </>
+  );
+}
+
+export default MenuMainButton;
