+++ src/global/global.css
@@ -17,18 +17,3 @@ body {
   background: var(--gray-900);
   color: var(--gray-300);
 }
-
-.container-fluid {
-  background-color: purple;
-}
-
-.container {
-  max-width: 98rem;
-  background-color: red;
-  margin: 0 auto;
-}
-
-.content {
-  background-color: blue;
-  margin: 3.2rem;
-}
